+++ chaoshoneycomb/__init__.py
@@ -80,4 +80,5 @@ def load_exported_activities() -> List[DiscoveredActivities]:
     activities.extend(discover_probes("chaoshoneycomb.slo.probes"))
     activities.extend(discover_probes("chaoshoneycomb.trigger.probes"))
     activities.extend(discover_actions("chaoshoneycomb.marker.actions"))
+    activities.extend(discover_probes("chaoshoneycomb.query.probes"))
     return activities